Assignment3/requirements.txt
Faker==0.9.2
mysql-connector-python==8.0.13
protobuf==3.6.1
python-dateutil==2.7.5
six==1.11.0
text-unidecode==1.2
